Fixed small issue with 29700144. This game had originally not contained a record for the game start action. I added it to the game's individual file, but it had the same ActionID as the tip off action, 1.
This fix adjusted the ActionIDs accordingly

Adding Periods to GameExt

Adding Version tracking functionality. Since i'll be releasing versions on GitHub, I want to make sure everyone's Toolbox will still work. When the Toolbox is opened and, only after a connection has been established and it is verified that the database already exists, it will verify that any changes in Database structure are accounted for. In this instance, I'm adding Periods to GameExt. If a user last opened Toolbox v1.0 and built their database with that, their GameExt table wouldn't have this column. If they download v1.01 and try to Populate or repair an already existing database, it would error out because v1.01 is now adding Periods. This tracking/check functionality will make sure the database is always up to date with whatever Toolbox version is being used.