Added fix for 21900884 when pulled via endpoints.
If we were to use the data found at this game's endpoint (https://cdn.nba.com/static/json/liveData/boxscore/boxscore_0021900884.json), we'd find that there are a lot of erroneous values.

For example, period = 0, attendance = 0, duration = 0. Further, for the actual Team Boxscores, all of the values are 0s.

The fix was to point this game to draw from the datafile in lieu of its endpoint