Enhanced PlayByPlay data's robustness and usefulness by populating ScoreHome and ScoreAway for every single record. Prior to doing so, the NBA's historic data would show those values as 0s or null if the action didn't result in any Points Generated.
Fixed issues with some inconsistencies in some 2019 games. It had the ShotResult value as ": Missed" and ": Made", which resulted in PointsGenerated not being populated in some instances. Fixed all related issues here

Fixed issue in which Toolbox would erroneously display "Database update to v1.01 failed!" after building database for first time

Fixed issue with 21700025 PlayByPlay in which the data published by the NBA was incorrect.

Fixed issue with 41900111 where ActionID 366/ActionNumber 507 was duplicated in the data the NBA published. One of the rows didn't contain the assistPersonId value, so i skipped that less informative row and only brought 1 in.